chore(storybook-addon): use workspace plugin loader for buildless DX
