--- packages/react-components/react-storybook-addon/preset.dev.js
@@ -1,9 +1,0 @@
-function config(entry = []) {
-  return [...entry, require.resolve('./src/preset/preview')];
-}
-
-function managerEntries(entry = []) {
-  return [...entry, require.resolve('./src/preset/manager')];
-}
-
-module.exports = { managerEntries, config };
