+++ packages/react-components/react-storybook-addon/preset.js
@@ -1,9 +1,9 @@
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./preset.prod');
-} else {
-  const { workspaceRoot } = require('nx/src/utils/app-root');
-  const { registerTsProject } = require('nx/src/utils/register');
-
-  registerTsProject(workspaceRoot, 'tsconfig.base.json');
-  module.exports = require('./preset.dev');
+function config(entry = []) {
+  return [...entry, require.resolve('./lib/preset/preview')];
 }
+
+function managerEntries(entry = []) {
+  return [...entry, require.resolve('./lib/preset/manager')];
+}
+
+module.exports = { managerEntries, config };
