--- packages/react-components/react-storybook-addon/preset.prod.js
@@ -1,9 +1,0 @@
-function config(entry = []) {
-  return [...entry, require.resolve('./lib/preset/preview')];
-}
-
-function managerEntries(entry = []) {
-  return [...entry, require.resolve('./lib/preset/manager')];
-}
-
-module.exports = { managerEntries, config };
